flexible-rag/requirements.txt
lancedb>=0.9.0
pyarrow
pandas
numpy
scikit-learn
rank-bm25
sentence-transformers
transformers
fastapi
uvicorn
pydantic
rich
tqdm
httpx
tenacity
python-dotenv
